requirements.txt
python-dotenv
python-multipart
fastapi
uvicorn
python-dotenv
requests
celery
redis
